backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.17.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
]
